@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.11.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -33,9 +33,9 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.2
+  cupertino_icons: ^1.0.4
   flutter_staggered_animations: ^1.0.0
-  group_button: ^3.3.1
+  group_button: ^4.5.0
   http: ^0.13.4
   favorite_button: ^0.0.4
   bottom_navy_bar: ^6.0.0
@@ -45,8 +45,10 @@ dependencies:
   flutter_form_builder: ^7.1.1
   get: ^4.6.1
   wc_form_validators: ^1.0.0
-
+  json_annotation: ^4.4.0
 dev_dependencies:
+  build_runner: ^2.1.7
+  json_serializable: ^6.1.4
   flutter_test:
     sdk: flutter
 
@@ -55,7 +57,7 @@ dev_dependencies:
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^1.0.0
+  flutter_lints: ^1.0.4
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
