@@ -5,7 +5,7 @@ import 'package:bottom_navy_bar/bottom_navy_bar.dart';
 // Custom class to show the data.
 class Category {
   String name, image;
-  Category({ this.name,  this.image});
+  Category({ required this.name,  required this.image});
 }
 
 // Custom list to show the data.
